Move contact form into component
So it can be re-used on other pages
