+++ src/ui/components/FormContact/stylesheet.css
@@ -1,0 +1,81 @@
+@block screen from "../../styles/blocks/screen.block.css";
+@block button from "../../styles/blocks/button.block.css";
+@block typography from "../../styles/blocks/typography.block.css";
+
+:scope {
+  block-name: FormContact;
+  extends: screen;
+}
+
+:scope[state|form='submitting'] {
+}
+
+.spinner {
+  display: none;
+}
+
+.fieldset {
+  position: relative;
+}
+
+.legend {
+  position: absolute;
+  height: 0;
+  overflow: hidden;
+}
+
+.item,
+.group {
+  margin-bottom: 3rem;
+}
+
+.label,
+.control-label {
+  display: block;
+  margin-bottom: 0.5rem;
+  font-size: 2.625rem;
+  line-height: 1.523809524;
+  color: #5b5b5b;
+}
+
+.control {
+  width: 100%;
+  padding: calc(1.5rem - 1px) calc(1.5rem - 1px);
+  border: 1px solid #ccc;
+  box-sizing: border-box;
+  font-size: 2rem;
+  line-height: 3rem;
+}
+
+.control:focus {
+  outline: 0;
+  box-shadow: 0 0 3px 0 var(--color-accent);
+}
+
+.textarea {
+  width: 100%;
+  height: 4rem;
+  border: 1px solid #ccc;
+}
+
+:scope[state|form='submitting'] .item {
+  opacity: 0.35;
+}
+
+:scope[state|form='submitting'] .spinner {
+  display: block;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%) rotate(0);
+  animation: rotate 0.6s infinite linear;
+}
+
+@keyframes rotate {
+  0% {
+    transform: translate(-50%, -50%) rotate(0);
+  }
+  100% {
+    transform: translate(-50%, -50%) rotate(360deg);
+  }
+}
