+++ src/ui/components/FormContact/template.hbs
@@ -1,0 +1,33 @@
+<form onsubmit={{action this.submit}} state:form={{this.formState}} disabled={{this.isSubmitting}}>
+  <fieldset class="fieldset">
+    <legend class="legend">Contact</legend>
+    <svg class="spinner" width="58" height="58" xmlns="http://www.w3.org/2000/svg"><g transform="translate(2 2)" stroke-width="3" fill="none" fill-rule="evenodd"><circle stroke="#B3D9FF" cx="27" cy="27" r="27"/><path d="M54 27C54 12.09 41.91 0 27 0" stroke="#007DF6"/></g></svg>
+    <p class="item">
+      <label for="name" class="label">
+        Name
+      </label>
+      <input class="control" id="name" type="text" autocomplete="name" disabled={{this.isSubmitting}} />
+    </p>
+    <p class="item">
+      <label for="email" class="label">
+        E-Mail
+      </label>
+      <input class="control" id="email" name="email" autocomplete="email" disabled={{this.isSubmitting}} />
+    </p>
+    <p class="item">
+      <label for="message" class="label">
+        Tell us about your project
+      </label>
+      <textarea class="control" id="message" cols="30" rows="10" disabled={{this.isSubmitting}}></textarea>
+    </p>
+  </fieldset>
+  <p>
+    <button type="submit" class="button.button">
+      {{#if this.isSubmitting}}
+        Sending…
+      {{else}}
+        Send message
+      {{/if}}
+    </button>
+  </p>
+</form>
