+++ src/ui/components/PageContact/stylesheet.css
@@ -2,8 +2,6 @@
 @block main-block from "../../styles/blocks/main-block.block.css";
 @block header-block from "../../styles/blocks/header-block.block.css";
 @block typography from "../../styles/blocks/typography.block.css";
-@block form from "../../styles/blocks/form.block.css";
-@block button from "../../styles/blocks/button.block.css";
 
 :scope {
   block-name: PageContact;
