+++ src/ui/components/PageTraining/stylesheet.css
@@ -6,8 +6,6 @@
 @block shape-left-md from "../../styles/blocks/shape/shape-left-md.block.css";
 @block fluid-image from "../../styles/blocks/fluid-image.block.css";
 @block card from "../../styles/blocks/card.block.css";
-@block form from "../../styles/blocks/form.block.css";
-@block button from "../../styles/blocks/button.block.css";
 
 :scope {
   block-name: PageTraining;
