+++ src/ui/components/PageTraining/template.hbs
@@ -30,31 +30,7 @@
           <h3 class="typography.headline-3">
             Partner with us to help lorem ipsum sit amet
           </h3>
-          <form class="form">
-            <div class="form.group">
-              <label class="form.control-label">
-                Your Name
-              </label>
-              <input class="form.control" type="text" name="" />
-            </div>
-            <div class="form.group">
-              <label class="form.control-label">
-                Your Email
-              </label>
-              <input class="form.control" type="text" name="" />
-            </div>
-            <div class="form.group">
-              <label class="form.control-label">
-                Tell us about your project
-              </label>
-              <textarea class="form.textarea"></textarea>
-            </div>
-            <div class="form.group">
-              <button class="button.button">
-                Submit
-              </button>
-            </div>
-          </form>
+          <FormContact />
         </div>
       </div>
     </div>
