--- src/ui/styles/blocks/form.block.css
@@ -1,30 +1,0 @@
-:scope {
-  block-name: form;
-}
-
-.item,
-.group {
-  margin-bottom: 3rem;
-}
-
-.label,
-.control-label {
-  display: block;
-  margin-bottom: 0.5rem;
-  font-size: 2.625rem;
-  line-height: 1.523809524;
-  color: #5b5b5b;
-}
-
-.control {
-  width: 100%;
-  padding: 1rem;
-  border: 1px solid #ccc;
-  box-sizing: border-box;
-}
-
-.textarea {
-  width: 100%;
-  height: 4rem;
-  border: 1px solid #ccc;
-}
